Drastically simplified types and API, removed function type declarations. Two new functions: attestation and assertion, signature not final. Removed utils reexport, commonjs fixes
