--- cjs/index.d.ts
@@ -1,3 +1,0 @@
-export * from "./utils.js";
-export * from "./verify.js";
-export * from "./parse.js";
